Services/Mikhail/pyproject.toml
[project]
name = "mikhail"
version = "0.1.0"
description = "Backend service for Kingdom-System"
authors = [
    {name = "TwelveFacedJanus"}
]
license = {text = "BSD-3 Clause"}
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv (>=0.9.9,<0.10.0)",
    "grpcio-tools (>=1.72.1,<2.0.0)",
    "grpcio (>=1.72.1,<2.0.0)",
    "requests (>=2.31.0,<3.0.0)",
    "python-jose[cryptography] (>=3.3.0,<4.0.0)",
    "google-auth (>=2.27.0,<3.0.0)",
    "redis (>=5.0.1,<6.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
